Fixed Main Menu Selection Issues;
Bounding boxes for mouse selection are no longer messed up. Controller selection no longer breaks as soon as the player tries to select anything with their mouse. This only applies to the primary 3 buttons. Will bring changes to other sections of the menu and to the pause menu in a later commit.